requirements.txt
flask==2.3.3
python-dotenv==1.1.0
google-adk==0.3.0
google-generativeai==0.8.5
SQLAlchemy==2.0.40
aiohttp==3.11.18
pydantic==2.11.4
gunicorn==21.2.0
requests==2.32.3
mysql-connector-python>=8.0.33 
requests>=2.32.3
litellm>=1.40.0 
schedule>=1.2.1
asgiref
uvicorn
deprecated
pytz>=2023.3
litellm
opensearch-py